deduplicate/unify engine launching logic
