+++ vllm/entrypoints/cli/serve.py
@@ -8,7 +8,6 @@ import sys
 from typing import Optional
 
 import uvloop
-import zmq
 
 import vllm
 import vllm.envs as envs
@@ -22,18 +21,13 @@ from vllm.entrypoints.utils import (VLLM_SUBCMD_PARSER_EPILOG,
 from vllm.executor.multiproc_worker_utils import _add_prefix
 from vllm.logger import init_logger
 from vllm.usage.usage_lib import UsageContext
-from vllm.utils import (FlexibleArgumentParser, get_open_zmq_ipc_path,
-                        get_tcp_uri, zmq_socket_ctx)
-from vllm.v1.engine.coordinator import DPCoordinator
+from vllm.utils import FlexibleArgumentParser, get_tcp_uri
 from vllm.v1.engine.core import EngineCoreProc
-from vllm.v1.engine.core_client import CoreEngineProcManager
+from vllm.v1.engine.utils import (APIServerProcessManager,
+                                  CoreEngineProcManager, launch_core_engines,
+                                  wait_for_completion_or_failure)
 from vllm.v1.executor.abstract import Executor
 from vllm.v1.metrics.prometheus import setup_multiprocess_prometheus
-from vllm.v1.utils import (APIServerProcessManager, CoreEngine,
-                           CoreEngineActorManager, EngineZmqAddresses,
-                           get_engine_client_zmq_addr,
-                           wait_for_completion_or_failure,
-                           wait_for_engine_startup)
 
 logger = init_logger(__name__)
 
@@ -194,116 +188,35 @@ def run_multi_api_server(args: argparse.Namespace):
                 " api_server_count > 1")
             model_config.disable_mm_preprocessor_cache = True
 
+    executor_class = Executor.get_class(vllm_config)
+    log_stats = not engine_args.disable_log_stats
+
     parallel_config = vllm_config.parallel_config
-
-    dp_size = parallel_config.data_parallel_size
     dp_rank = parallel_config.data_parallel_rank
-    local_engine_count = parallel_config.data_parallel_size_local
-    host = parallel_config.data_parallel_master_ip
     external_dp_lb = parallel_config.data_parallel_external_lb
-
-    # The handshake socket can be IPC if this front-end is colocated
-    # with all engines.
-    handshake_local_only = local_engine_count == dp_size
-
-    # client_local_only = True for cases where this front-end
-    # sends requests only to colocated engines.
-    client_local_only = handshake_local_only or external_dp_lb
-
     assert external_dp_lb or dp_rank == 0
 
-    # Set up input and output addresses.
-    input_addresses = [
-        get_engine_client_zmq_addr(client_local_only, host)
-        for _ in range(num_api_servers)
-    ]
-    output_addresses = [
-        get_engine_client_zmq_addr(client_local_only, host)
-        for _ in range(num_api_servers)
-    ]
-
-    addresses = EngineZmqAddresses(
-        inputs=input_addresses,
-        outputs=output_addresses,
-    )
-
-    # Set up coordinator for dp > 1.
-    # This runs only with the rank 0 engine, or when not colocated with any
-    # engines (in which case the value of dp_rank here will also be 0).
-    coordinator = None
-    stats_update_address = None
-    if dp_size > 1 and dp_rank == 0:
-        coordinator = DPCoordinator(parallel_config)
-        addresses.coordinator_input, addresses.coordinator_output = (
-            coordinator.get_engine_socket_addresses())
-        addresses.frontend_stats_publish_address = (
-            coordinator.get_stats_publish_address())
-        stats_update_address = coordinator.get_stats_publish_address()
-        logger.info("Started DP Coordinator process (PID: %d)",
-                    coordinator.proc.pid)
-
     api_server_manager: Optional[APIServerProcessManager] = None
-    # Construct common args for the APIServerProcessManager up-front.
-    api_server_manager_kwargs = dict(
-        target_server_fn=run_api_server_worker_proc,
-        listen_address=listen_address,
-        sock=sock,
-        args=args,
-        num_servers=num_api_servers,
-        input_addresses=input_addresses,
-        output_addresses=output_addresses,
-        stats_update_address=stats_update_address)
 
-    if parallel_config.data_parallel_backend == "ray":
-        logger.info("Starting ray-based data parallel backend")
+    with launch_core_engines(vllm_config, executor_class, log_stats,
+                             num_api_servers) as (local_engine_manager,
+                                                  coordinator, addresses):
 
-        engine_actor_manager = CoreEngineActorManager(
-            vllm_config=vllm_config,
-            addresses=addresses,
-            executor_class=Executor.get_class(vllm_config),
-            log_stats=not engine_args.disable_log_stats,
-        )
-        # Start API servers using the manager
-        api_server_manager = APIServerProcessManager(
-            **api_server_manager_kwargs)
-
-        wait_for_completion_or_failure(api_server_manager=api_server_manager,
-                                       engine_manager=engine_actor_manager,
-                                       coordinator=coordinator)
-        return
-
-    handshake_address = get_engine_client_zmq_addr(
-        handshake_local_only, host, parallel_config.data_parallel_rpc_port)
-
-    if external_dp_lb and dp_rank > 0:
-        assert not handshake_local_only
-        local_handshake_address = get_open_zmq_ipc_path()
-        client_handshake_address = local_handshake_address
-    else:
-        local_handshake_address = handshake_address
-        client_handshake_address = None
-
-    with zmq_socket_ctx(local_handshake_address, zmq.ROUTER,
-                        bind=True) as handshake_socket:
-
-        # Start local engines.
-        if not local_engine_count:
-            local_engine_manager = None
-        else:
-            local_engine_manager = CoreEngineProcManager(
-                EngineCoreProc.run_engine_core,
-                vllm_config=vllm_config,
-                executor_class=Executor.get_class(vllm_config),
-                log_stats=not engine_args.disable_log_stats,
-                handshake_address=handshake_address,
-                client_handshake_address=client_handshake_address,
-                local_client=True,
-                local_engine_count=local_engine_count,
-                start_index=dp_rank,
-                local_start_index=0)
+        # Construct common args for the APIServerProcessManager up-front.
+        api_server_manager_kwargs = dict(
+            target_server_fn=run_api_server_worker_proc,
+            listen_address=listen_address,
+            sock=sock,
+            args=args,
+            num_servers=num_api_servers,
+            input_addresses=addresses.inputs,
+            output_addresses=addresses.outputs,
+            stats_update_address=coordinator.get_stats_publish_address()
+            if coordinator else None)
 
         # For dp ranks > 0 in external DP LB mode, we must delay the
-        # start of the API servers until the local engine is started,
+        # start of the API servers until the local engine is started
+        # (after the launcher context manager exits),
         # since we get the front-end stats update address from the coordinator
         # via the handshake with the local engine.
         if dp_rank == 0 or not external_dp_lb:
@@ -311,37 +224,17 @@ def run_multi_api_server(args: argparse.Namespace):
             api_server_manager = APIServerProcessManager(
                 **api_server_manager_kwargs)
 
-        # Wait for engine handshakes to complete.
-        if external_dp_lb and dp_rank > 0:
-            assert local_engine_count == 1
-            core_engines = [CoreEngine(index=dp_rank, local=True)]
-        else:
-            core_engines = [
-                CoreEngine(index=i, local=(i < local_engine_count))
-                for i in range(dp_size)
-            ]
+    # Start API servers now if they weren't already started.
+    if api_server_manager is None:
+        api_server_manager_kwargs["stats_update_address"] = (
+            addresses.frontend_stats_publish_address)
+        api_server_manager = APIServerProcessManager(
+            **api_server_manager_kwargs)
 
-        wait_for_engine_startup(
-            handshake_socket,
-            addresses,
-            core_engines,
-            parallel_config,
-            vllm_config.cache_config,
-            local_engine_manager,
-            coordinator.proc if coordinator else None,
-        )
-
-        # Start API servers now if they weren't already started.
-        if api_server_manager is None:
-            api_server_manager_kwargs["stats_update_address"] = (
-                addresses.frontend_stats_publish_address)
-            api_server_manager = APIServerProcessManager(
-                **api_server_manager_kwargs)
-
-        # Wait for API servers
-        wait_for_completion_or_failure(api_server_manager=api_server_manager,
-                                       engine_manager=local_engine_manager,
-                                       coordinator=coordinator)
+    # Wait for API servers
+    wait_for_completion_or_failure(api_server_manager=api_server_manager,
+                                   engine_manager=local_engine_manager,
+                                   coordinator=coordinator)
 
 
 def run_api_server_worker_proc(listen_address,
